Added s3_bucket model
